refactored product query to be part of the app.tsx
