+++ client/src/pages/home.tsx
@@ -5,30 +5,14 @@ import { CartItem } from "./cart";
 
 type HomePageProps = {
     setCart: React.Dispatch<React.SetStateAction<CartItem[]>>;
+    products: Product[]
 };
 
-const HomePage = ({setCart}: HomePageProps) => {
-    const [products, setProducts] = useState<Product[]>([]);
-    const productQuery = useQuery({
-        queryKey: ['products'], 
-        queryFn: async () => {
-            const response = await fetch(`${import.meta.env.VITE_API_URL}/products`);
-            const data = await response.json();
-            return data;
-        }});
-
-    useEffect(() => {
-        if(productQuery.isSuccess) {
-            setProducts(productQuery.data.products);
-        }
-    }, [productQuery]);
+const HomePage = ({setCart, products}: HomePageProps) => {
 
     const addToCart = (cartProduct: CartItem) => {
         setCart((prevCart) => [...prevCart, cartProduct]);
     };
-
-    if(productQuery.isLoading) return <div>Loading...</div>;
-    if(productQuery.isError) return <div>Error fetching products</div>;
 
     return (
         <div className="flex justify-center items-center flex-col w-full gap-2 md:gap-0">
